Refactor namespaces for version 1 of the Mattermost REST api
