+++ OutlookMatters.Core.Test/Mattermost/v1/ChatPostImplTest.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using Moq;
 using NUnit.Framework;
-using OutlookMatters.Core.Mattermost;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 
-namespace Test.OutlookMatters.Core.Mattermost
+namespace Test.OutlookMatters.Core.Mattermost.v1
 {
     [TestFixture]
     public class ChatPostImplTest
