+++ OutlookMatters.Core.Test/Mattermost/v1/HttpClientTest.cs
@@ -3,10 +3,10 @@ using FluentAssertions;
 using Moq;
 using NUnit.Framework;
 using OutlookMatters.Core.Chat;
-using OutlookMatters.Core.Mattermost;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 
-namespace Test.OutlookMatters.Core.Mattermost
+namespace Test.OutlookMatters.Core.Mattermost.v1
 {
     [TestFixture]
     public class HttpClientTest
