+++ OutlookMatters.Core.Test/Mattermost/v1/Interface/MattermostExceptionTest.cs
@@ -1,9 +1,8 @@
 ﻿using FluentAssertions;
 using NUnit.Framework;
-using OutlookMatters.Core.Chat;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 
-namespace Test.OutlookMatters.Core.Mattermost.Interface
+namespace Test.OutlookMatters.Core.Mattermost.v1.Interface
 {
     [TestFixture]
     public class MattermostExceptionTest
