+++ OutlookMatters.Core.Test/Session/SingleSignOnSessionRepositoryTest.cs
@@ -3,7 +3,6 @@ using FluentAssertions;
 using Moq;
 using NUnit.Framework;
 using OutlookMatters.Core.Chat;
-using OutlookMatters.Core.Mattermost.Interface;
 using OutlookMatters.Core.Security;
 using OutlookMatters.Core.Session;
 using OutlookMatters.Core.Settings;
