+++ OutlookMatters.Core/ContextMenu/MailItemContextMenuEntry.cs
@@ -8,7 +8,7 @@ using Newtonsoft.Json;
 using OutlookMatters.Core.Chat;
 using OutlookMatters.Core.Error;
 using OutlookMatters.Core.Mail;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 using OutlookMatters.Core.Reply;
 using OutlookMatters.Core.Session;
 using OutlookMatters.Core.Settings;
