+++ OutlookMatters.Core/Error/IErrorDisplay.cs
@@ -1,6 +1,6 @@
 using System;
 using OutlookMatters.Core.Chat;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 
 namespace OutlookMatters.Core.Error
 {
