+++ OutlookMatters.Core/Mattermost/v1/HttpSession.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.Linq;
 using OutlookMatters.Core.Chat;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 
-namespace OutlookMatters.Core.Mattermost
+namespace OutlookMatters.Core.Mattermost.v1
 {
     public class HttpSession : ISession
     {
