+++ OutlookMatters.Core/Mattermost/v1/IRestService.cs
@@ -1,7 +1,7 @@
 ﻿using System;
-using OutlookMatters.Core.Mattermost.Interface;
+using OutlookMatters.Core.Mattermost.v1.Interface;
 
-namespace OutlookMatters.Core.Mattermost
+namespace OutlookMatters.Core.Mattermost.v1
 {
     public interface IRestService
     {
