+++ OutlookMatters.Core/Mattermost/v1/Interface/Error.cs
@@ -1,4 +1,4 @@
-﻿namespace OutlookMatters.Core.Mattermost.Interface
+﻿namespace OutlookMatters.Core.Mattermost.v1.Interface
 {
     public class Error
     {
