+++ OutlookMatters.Core/Mattermost/v1/Interface/MattermostException.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace OutlookMatters.Core.Mattermost.Interface
+namespace OutlookMatters.Core.Mattermost.v1.Interface
 {
     public class MattermostException : Exception
     {
