+++ OutlookMatters.Core/Mattermost/v1/Interface/Post.cs
@@ -1,6 +1,6 @@
 using Newtonsoft.Json;
 
-namespace OutlookMatters.Core.Mattermost.Interface
+namespace OutlookMatters.Core.Mattermost.v1.Interface
 {
     public class Post
     {
