+++ OutlookMatters.Core/OutlookMatters.Core.csproj
@@ -91,30 +91,30 @@
     <Compile Include="Http\IHttpClient.cs" />
     <Compile Include="Http\IHttpRequest.cs" />
     <Compile Include="Http\IHttpResponse.cs" />
-    <Compile Include="Mattermost\ChatChannelImpl.cs" />
-    <Compile Include="Mattermost\ChatPostImpl.cs" />
-    <Compile Include="Mattermost\HttpClient.cs" />
-    <Compile Include="Mattermost\HttpRestService.cs" />
-    <Compile Include="Mattermost\HttpSession.cs" />
-    <Compile Include="Mattermost\HttpSessionFactory.cs" />
-    <Compile Include="Mattermost\IChatChannelFactory.cs" />
-    <Compile Include="Mattermost\IChatPostFactory.cs" />
-    <Compile Include="Mattermost\IRestService.cs" />
-    <Compile Include="Mattermost\ISessionFactory.cs" />
+    <Compile Include="Mattermost\v1\ChatChannelImpl.cs" />
+    <Compile Include="Mattermost\v1\ChatPostImpl.cs" />
+    <Compile Include="Mattermost\v1\HttpClient.cs" />
+    <Compile Include="Mattermost\v1\HttpRestService.cs" />
+    <Compile Include="Mattermost\v1\HttpSession.cs" />
+    <Compile Include="Mattermost\v1\HttpSessionFactory.cs" />
+    <Compile Include="Mattermost\v1\IChatChannelFactory.cs" />
+    <Compile Include="Mattermost\v1\IChatPostFactory.cs" />
+    <Compile Include="Mattermost\v1\Interface\Channel.cs" />
+    <Compile Include="Mattermost\v1\Interface\ChannelList.cs" />
+    <Compile Include="Mattermost\v1\Interface\ChannelType.cs" />
+    <Compile Include="Mattermost\v1\Interface\Error.cs" />
+    <Compile Include="Mattermost\v1\Interface\Login.cs" />
+    <Compile Include="Mattermost\v1\Interface\MattermostException.cs" />
+    <Compile Include="Mattermost\v1\Interface\Post.cs" />
+    <Compile Include="Mattermost\v1\Interface\Thread.cs" />
+    <Compile Include="Mattermost\v1\Interface\User.cs" />
+    <Compile Include="Mattermost\v1\IRestService.cs" />
+    <Compile Include="Mattermost\v1\ISessionFactory.cs" />
     <Compile Include="Session\ISessionRepository.cs" />
     <Compile Include="Mail\IExplorerService.cs" />
     <Compile Include="Mail\IMailExplorer.cs" />
     <Compile Include="Mail\MailException.cs" />
     <Compile Include="Mail\OutlookMailExplorer.cs" />
-    <Compile Include="Mattermost\Interface\Channel.cs" />
-    <Compile Include="Mattermost\Interface\ChannelList.cs" />
-    <Compile Include="Mattermost\Interface\ChannelType.cs" />
-    <Compile Include="Mattermost\Interface\Error.cs" />
-    <Compile Include="Mattermost\Interface\Login.cs" />
-    <Compile Include="Mattermost\Interface\Post.cs" />
-    <Compile Include="Mattermost\Interface\Thread.cs" />
-    <Compile Include="Mattermost\Interface\User.cs" />
-    <Compile Include="Mattermost\Interface\MattermostException.cs" />
     <Compile Include="Session\SingleSignOnSessionRepository.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Reply\IStringProvider.cs" />
