+++ OutlookMatters.Core/Session/ISessionRepository.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using OutlookMatters.Core.Chat;
-using OutlookMatters.Core.Mattermost.Interface;
 
 namespace OutlookMatters.Core.Session
 {
